requirements.txt
streamlit
tensorflow == 2.15.1
Pillow
numpy